reworked geometry/blob casts
