+++ spatial/include/spatial/core/functions/cast.hpp
@@ -11,11 +11,13 @@ public:
 		RegisterVarcharCasts(context);
 		RegisterDimensionalCasts(context);
 		RegisterGeometryCasts(context);
+		RegisterWKBCasts(context);
 	}
 private:
 	static void RegisterVarcharCasts(ClientContext &context);
 	static void RegisterDimensionalCasts(ClientContext &context);
 	static void RegisterGeometryCasts(ClientContext &context);
+	static void RegisterWKBCasts(ClientContext &context);
 };
 
 } // namespace core
