+++ spatial/src/spatial/core/functions/cast/CMakeLists.txt
@@ -3,5 +3,6 @@ set(EXTENSION_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/dimensional_cast.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/geometry_cast.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/varchar_cast.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/wkb_cast.cpp
     PARENT_SCOPE
 )
